Moving annotations info their own package
