+++ src/main/java/org/dataloader/BatchLoaderWithContext.java
@@ -1,5 +1,7 @@
 package org.dataloader;
 
+import org.dataloader.annotations.PublicSpi;
+
 import java.util.List;
 import java.util.concurrent.CompletionStage;
